test: fold MC search helpers into existing canonical files
Per PR #1048 reviewer pattern (Julien's 484ea09b3 "delete
tests/common/multicluster_search and pre-stage proxy_svc_fqdn"),
no new helper file. Distribute the four helpers to their
existing canonical homes:

- MCSearchDeploymentHelper -> tests/common/search/search_deployment_helper.py
  (peers with SearchDeploymentHelper; both own search-fixture lifecycle).
  proxy_svc_fqdn() now delegates to the existing
  search_resource_names.mc_proxy_svc_fqdn (no duplicated FQDN logic).
- replicate_secret + assert_resource_in_cluster + assert_deployment_ready_in_cluster
  -> tests/common/multicluster/multicluster_utils.py (generic MC test
  utilities, none are search-specific).

Deletes tests/common/search/multicluster_search_helper.py from the
previous commit (consolidation done in the wrong shape).
